requirements.txt
streamlit==1.7.0
numpy==1.22.3
pandas==1.2.4
Pillow==8.2.0